+++ app/views/spina/admin/shared/_navigation.html.erb
@@ -40,7 +40,10 @@
         <% nav.links do %>
           <%= render Spina::MainNavigation::LinkComponent.new("General", spina.edit_admin_account_path, active: controller_name == "accounts") %>
           <%= render Spina::MainNavigation::LinkComponent.new(t('spina.website.theme'), spina.edit_admin_theme_path, active: controller_name == "theme") %>
-          <%= render Spina::MainNavigation::LinkComponent.new("Users", spina.admin_users_path, active: controller_name == "users") %>
+          
+          <% if Spina.config.authentication == "Spina::Auth" %>
+            <%= render Spina::MainNavigation::LinkComponent.new("Users", spina.admin_users_path, active: controller_name == "users") %>
+          <% end %>
           
           <%= render Spina::Hooks::HookComponent.new(partial: "settings_secondary_navigation") %>
